completed most of first page
