+++ app/component-data/Articles/index.tsx
@@ -1,0 +1,14 @@
+import { FeaturedArticleCardProps } from "~/ui/design-system/src/lib/Components/FeaturedArticleCard"
+
+const introToFlowBlockchain: FeaturedArticleCardProps = {
+  heading: "Introduction to Flow blockchain",
+  tags: ["protocol", "network"],
+  description: `When Dapper Labs built Crypto Kitties we learned a lot.
+    Most importantly, we realized that the technology at the time was not ready for this kind of application.
+    Being the visionaries we are, we set to build a better tech for what we plan to do.
+    We set to build what is now Flow blockchain.`,
+  link: "https://jan-bernatik.medium.com/introduction-to-flow-blockchain-7532977c8af8",
+  ctaText: "View Article",
+}
+
+export { introToFlowBlockchain }
